requirements.txt
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
tzdata; platform_system == "Windows"